back - feat: add find order logic to deleteOrder useCase

+++ back-end/src/modules/orders/services/deleteOrder/DeleteOrderUseCase.ts
@@ -2,6 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { OrderDTO } from "../../DTO/OrderDTO";
 import { IOrderRepository } from "../../database/interface/IOrdersRepository";
 import { Order } from "../../entities/Order";
+import { OrderNotExists } from "../../errors/OrderNotExists";
 
 
 @Injectable()
@@ -12,7 +13,10 @@ export class DeleteOrderUseCase {
 
   async execute(order_id: string): Promise<void> {
 
-    
+    const order = await this.orderRepository.findById(order_id)
+    if(!order) {
+      throw new OrderNotExists();
+    }
   
     await this.orderRepository.deleteOrder(order_id)
